Move _options to snapcraft/project
