+++ snapcraft/internal/project_loader/_config.py
@@ -26,8 +26,10 @@ import yaml.reader
 from typing import Set  # noqa
 
 
-import snapcraft
+from snapcraft import project
+from snapcraft.project._project_info import ProjectInfo
 from snapcraft.internal import deprecations, remote_parts, states
+
 from ._schema import Validator
 from ._parts_config import PartsConfig
 from ._env import (
@@ -85,9 +87,9 @@ class Config:
             self._remote_parts_attr = remote_parts.get_remote_parts()
         return self._remote_parts_attr
 
-    def __init__(self, project_options: snapcraft.Project=None) -> None:
+    def __init__(self, project_options: project.Project=None) -> None:
         if project_options is None:
-            project_options = snapcraft.Project()
+            project_options = project.Project()
 
         self.build_snaps = set()  # type: Set[str]
         self.build_tools = set()  # type: Set[str]
@@ -109,7 +111,7 @@ class Config:
         self.data = self._expand_env(snapcraft_yaml)
         # We need to set the ProjectInfo here because ProjectOptions is
         # created in the CLI.
-        self._project_options.info = snapcraft.ProjectInfo(self.data)
+        self._project_options.info = ProjectInfo(self.data)
         self._ensure_no_duplicate_app_aliases()
 
         grammar_processor = grammar_processing.GlobalGrammarProcessor(
