+++ snapcraft/project/_project_options.py
@@ -17,128 +17,25 @@
 import logging
 import multiprocessing
 import os
-import platform
-import sys
 from contextlib import suppress
-from typing import Any, Dict, List, Set, Union  # noqa: F401
+from typing import List, Set, Union  # noqa: F401
 
 from snapcraft import file_utils
 from snapcraft.internal import common, errors, os_release
 from snapcraft.internal.deprecations import handle_deprecation_notice
 
 
+from ._arch_mapping import (
+    _ARCH_TRANSLATIONS,
+    _HOST_CODENAME_FOR_BASE,
+    _HOST_COMPATIBILITY,
+    _LINKER_VERSION_FOR_BASE,
+    _get_platform_architecture,
+    _find_machine
+)
+
+
 logger = logging.getLogger(__name__)
-
-
-_ARCH_TRANSLATIONS = {
-    'armv7l': {
-        'kernel': 'arm',
-        'deb': 'armhf',
-        'uts_machine': 'arm',
-        'cross-compiler-prefix': 'arm-linux-gnueabihf-',
-        'cross-build-packages': ['gcc-arm-linux-gnueabihf',
-                                 'libc6-dev-armhf-cross'],
-        'triplet': 'arm-linux-gnueabihf',
-        'core-dynamic-linker': 'lib/ld-linux-armhf.so.3',
-    },
-    'aarch64': {
-        'kernel': 'arm64',
-        'deb': 'arm64',
-        'uts_machine': 'aarch64',
-        'cross-compiler-prefix': 'aarch64-linux-gnu-',
-        'cross-build-packages': ['gcc-aarch64-linux-gnu',
-                                 'libc6-dev-arm64-cross'],
-        'triplet': 'aarch64-linux-gnu',
-        'core-dynamic-linker': 'lib/ld-linux-aarch64.so.1',
-    },
-    'i686': {
-        'kernel': 'x86',
-        'deb': 'i386',
-        'uts_machine': 'i686',
-        'triplet': 'i386-linux-gnu',
-    },
-    'ppc64le': {
-        'kernel': 'powerpc',
-        'deb': 'ppc64el',
-        'uts_machine': 'ppc64el',
-        'cross-compiler-prefix': 'powerpc64le-linux-gnu-',
-        'cross-build-packages': ['gcc-powerpc64le-linux-gnu',
-                                 'libc6-dev-ppc64el-cross'],
-        'triplet': 'powerpc64le-linux-gnu',
-        'core-dynamic-linker': 'lib64/ld64.so.2',
-    },
-    'ppc': {
-        'kernel': 'powerpc',
-        'deb': 'powerpc',
-        'uts_machine': 'powerpc',
-        'cross-compiler-prefix': 'powerpc-linux-gnu-',
-        'cross-build-packages': ['gcc-powerpc-linux-gnu',
-                                 'libc6-dev-powerpc-cross'],
-        'triplet': 'powerpc-linux-gnu',
-    },
-    'x86_64': {
-        'kernel': 'x86',
-        'deb': 'amd64',
-        'uts_machine': 'x86_64',
-        'triplet': 'x86_64-linux-gnu',
-        'core-dynamic-linker': 'lib64/ld-linux-x86-64.so.2',
-    },
-    's390x': {
-        'kernel': 's390',
-        'deb': 's390x',
-        'uts_machine': 's390x',
-        'cross-compiler-prefix': 's390x-linux-gnu-',
-        'cross-build-packages': ['gcc-s390x-linux-gnu',
-                                 'libc6-dev-s390x-cross'],
-        'triplet': 's390x-linux-gnu',
-        'core-dynamic-linker': 'lib/ld64.so.1',
-    }
-}
-
-
-_32BIT_USERSPACE_ARCHITECTURE = {
-    'aarch64': 'armv7l',
-    'armv8l': 'armv7l',
-    'ppc64le': 'ppc',
-    'x86_64': 'i686',
-}
-
-
-_WINDOWS_TRANSLATIONS = {
-    'AMD64': 'x86_64'
-}
-
-
-_HOST_CODENAME_FOR_BASE = {
-    'core18': 'bionic',
-    'core': 'xenial',
-}
-_HOST_COMPATIBILITY = {
-    'xenial': ['trusty', 'xenial'],
-    'bionic': ['trusty', 'xenial', 'bionic'],
-}
-
-
-_LINKER_VERSION_FOR_BASE = {
-    'core18': '2.27',
-    'core': '2.23',
-}
-
-
-def _get_platform_architecture():
-    architecture = platform.machine()
-
-    # Translate the windows architectures we know of to architectures
-    # we can work with.
-    if sys.platform == 'win32':
-        architecture = _WINDOWS_TRANSLATIONS.get(architecture)
-
-    if platform.architecture()[0] == '32bit':
-        userspace = _32BIT_USERSPACE_ARCHITECTURE.get(architecture)
-        if userspace:
-            architecture = userspace
-
-    return architecture
 
 
 class ProjectOptions:
@@ -241,8 +138,6 @@ class ProjectOptions:
         self._set_machine(target_deb_arch)
         self.__debug = debug
 
-        super().__init__()
-
     def is_host_compatible_with_base(self, base: str) -> bool:
         """Determines if the host is compatible with the GLIBC of the base.
 
@@ -335,40 +230,3 @@ class ProjectOptions:
             logger.info('Setting target machine to {!r}'.format(
                 target_deb_arch))
         self.__machine_info = _ARCH_TRANSLATIONS[self.__target_machine]
-
-
-class ProjectInfo:
-    """Information gained from the snap's snapcraft.yaml file."""
-
-    def __init__(self, data: Dict[str, Any]) -> None:
-        self.name = data['name']
-        self.version = data['version']
-        self.summary = data.get('summary')
-        self.description = data.get('description')
-        self.confinement = data['confinement']
-
-
-class Project(ProjectOptions):
-    """All details around building a project concerning the build environment
-    and the snap being built."""
-
-    def __init__(self, *, use_geoip=False, parallel_builds=True,
-                 target_deb_arch: str=None, debug=False) -> None:
-        self.info = None  # type: ProjectInfo
-
-        super().__init__(use_geoip, parallel_builds, target_deb_arch, debug)
-
-
-def _get_deb_arch(machine):
-    return _ARCH_TRANSLATIONS[machine].get('deb', None)
-
-
-def _find_machine(deb_arch):
-    for machine in _ARCH_TRANSLATIONS:
-        if _ARCH_TRANSLATIONS[machine].get('deb', '') == deb_arch:
-            return machine
-        elif _ARCH_TRANSLATIONS[machine].get('uts_machine', '') == deb_arch:
-            return machine
-
-    raise errors.SnapcraftEnvironmentError(
-        'Cannot set machine from deb_arch {!r}'.format(deb_arch))
