+++ tests/fixture_setup.py
@@ -134,13 +134,14 @@ class FakeProjectOptions(fixtures.Fixture):
     def setUp(self):
         super().setUp()
 
-        patcher = mock.patch('snapcraft.Project')
+        patcher = mock.patch('snapcraft.project.Project')
         patcher.start()
         self.addCleanup(patcher.stop)
 
         # Special handling is required as ProjectOptions attributes are
         # handled with the @property decorator.
-        project_options_t = type(snapcraft.Project.return_value)
+        project_options_t = type(
+            snapcraft.project.Project.return_value)
         for key in self._kwargs:
             setattr(project_options_t, key, self._kwargs[key])
